# API/padelapp-api-model/src/main/resources/swagger/apiPadel2.yaml
openapi: 3.0.3
info:
  title: API Padelhg
  description: |
    This is a sample API for managing paddle court reservations, based on the OpenAPI 3.0 specification. This API allows performing operations related to paddle court reservations, such as creating, modifying, and canceling reservations.

  contact:
    email: ccoutomanul@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
  - url: http://localhost:8080
...
paths:
...
  /backoffice/booking:
    post:
      tags:
        - booking
        - backoffice
      summary: Add a new or array booking
      description: Add a new booking
      operationId: backOfficeAddBooking
      security:
        - bearerAuth: []
      requestBody:
        description: Create a new booking
        content:
          application/json:
            schema:
              type: object
              properties:
                bookings:
                  type: array
                  items:
                    $ref: '#/components/schemas/InputBookingDTO'
              required:
                - bookings
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingDTO'
        "405":
          description: Invalid input
...
  /backoffice/resource/{id}:
    put:
      tags:
        - resource
        - backoffice
      summary: Update an existing resource
      description: Update an existing resource by ID
      operationId: updateResource
      parameters:
        - name: id
          in: path
          description: ID of the resource to update
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      requestBody:
        description: Updated resource information
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InputResourceDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceDTO'
        "400":
          description: Invalid resource ID supplied
        "401":
          description: Unauthorized
        "404":
          description: Resource not found
        "405":
          description: Validation exception
    delete:
      tags:
        - resource
        - backoffice
      summary: Delete a resource by ID
      description: Delete an existing resource by ID
      operationId: deleteResource
      parameters:
        - name: id
          in: path
          description: ID of the resource to delete
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Resource deleted successfully
        "400":
          description: Invalid resource ID supplied
        "401":
          description: Unauthorized
        "404":
          description: Resource not found
  /backoffice/resource:
    post:
      tags:
        - resource
        - backoffice
      summary: create a new resourceDTO
      description: create a new resourceDTO
      operationId: createResourceDTO
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InputResourceDTO'
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceDTO'
        "405":
          description: Invalid input
...
  /backoffice/festive/{idfestive}:
    put:
      tags:
        - festive
        - backoffice
      summary: Update festive
      description: Change festive selected.
      operationId: updateFestive
      parameters:
        - name: idfestive
          in: path
          description: festive id to delete
          required: true
          schema:
            type: integer
      requestBody:
        description: Update an existent festive
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FestiveDTO'
      responses:
        default:
          description: successful operation
    delete:
      tags:
        - festive
        - backoffice
      summary: Delete festive
      description: This can only be done by the logged in user.
      operationId: deleteFestive
      parameters:
        - name: idfestive
          in: path
          description: The name that needs to be deleted
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted
        '400':
          description: Invalid username supplied
        '404':
          description: User not found
components:
  schemas:
...
    PlayerPosition:
      type: string
      description: player's favorite position
      enum:
        - DRIVE
        - REVERSE
        - INDIFFERENT
    ResourceDTOType:
      type: string
      description: booking type
      enum:
        - PADEL
        - ARENA
        - GYM
        - ECOWASH
    BookingType:
      type: string
      description: booking type
      enum:
        - TOURNAMENT
        - TRAINING
        - GAME
        - MATCH-TELEGRAM
        - LEAGUE
    BookingVisibility:
      type: string
      description: determines whether players can be joined or not
      enum:
        - PRIVATE
        - PUBLIC
    ResourceDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 2
        number:
          type: integer
          format: int32
          example: 12
        name:
          type: string
          example: Pista central
        resourceDTOType:
          $ref: '#/components/schemas/ResourceDTOType'
        timeSlot:
          type: integer
          description: reservable time slot
          example: 90
        startTimeSlot:
          type: string
          format: time
          example: 09:00:00
        endTimeSlot:
          type: string
          format: time
          example: 00:00:00
        basePrice:
          type: number
          format: double
          example: 5.5
        daysInAdvance:
          type: integer
          format: int32
          example: 3
    InputResourceDTO:
      type: object
      properties:
        number:
          type: integer
          format: int32
          example: 12
        name:
          type: string
          example: Pista central
        resourceDTOActivity:
          type: string
          example: Padel
        timeSlot:
          type: integer
          description: reservable time slot
          example: 90
        startTimeSlot:
          type: string
          format: time
          example: 09:00:00
        endTimeSlot:
          type: string
          format: time
          example: 00:00:00
        basePrice:
          type: number
          format: double
          example: 5.5
        daysInAdvance:
          type: integer
          format: int32
          example: 3
...
    UserBooking:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        username:
          type: string
          example: theUser
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: James
        email:
          type: string
          format: email
          example: john@email.com
        phone:
          type: string
          example: "12345"
        userPosition:
          $ref: '#/components/schemas/PlayerPosition'
...
    BookingDTO:
      type: object
      properties:
        idBooking:
          type: integer
          format: int64
          example: 10
        resourceDTO:
          $ref: '#/components/schemas/ResourceDTO'
        starDate:
          type: string
          format: date-time
          example: 2017-07-21T17:30:00Z
        endDate:
          type: string
          format: date-time
        ownerUser:
          $ref: '#/components/schemas/UserBooking'
        price:
          type: number
          format: double
        guessts:
          $ref: '#/components/schemas/UserBooking'
        bookingType:
          $ref: '#/components/schemas/BookingType'
        bookingVisibility:
          $ref: '#/components/schemas/BookingVisibility'
    InputBookingDTO:
      required:
        - endDate
        - resourceDTO
        - startDate
        - user
      type: object
      properties:
        resource:
          type: integer
          example: 6
        starDate:
          type: string
          format: date-time
          example: 2017-07-21T17:30:00Z
        endDate:
          type: string
          format: date-time
        user:
          type: integer
        bookingType:
          $ref: '#/components/schemas/BookingType'
        guessts:
          type: array
          items:
            type: string
            example: [juanca1223, manu34]
        bookingVisibility:
          $ref: '#/components/schemas/BookingVisibility'
        level:
          type: integer
          example: 5
    FestiveDTO:
      type: object
      properties:
        festiveId:
          type: integer
          format: int64
          example: 10
        date:
          type: string
          format: date-time
          example: 2023-03-28
        description:
          type: string
          example: Reconquista de Vigo
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    api_key:
      type: apiKey
      name: api_key
      in: header